Revert "Adding highlights section & refactoring utils to utils/formatting (#38)"
This reverts commit 9c9a0fcc393422b8ecc7ed1a045f31c261002790.
