+++ src/components/Preview.js
@@ -3,12 +3,6 @@ import marked from 'marked';
 import moment from 'moment';
 import { Dimmer, Loader } from 'semantic-ui-react';
 
-import {
-  formatSectionHeader,
-  formatHighlights,
-  formatItem,
-  replaceWikiLinks,
-} from 'utils/formatting/markdown';
 import { updateCard } from 'services/api';
 import Button from './inputs/Button';
 import Modal from './layout/Modal';
@@ -20,6 +14,46 @@ marked.setOptions({
   smartypants: false,
 });
 
+const getWikiLink = (code) => {
+  const linkBaseUrl = 'https://github.com/rangle/hub/wiki/';
+  const parts = code.split('|');
+  const visibleText = parts[0];
+  const link = (parts[1] || parts[0]).replace(/\s/g, '-');
+  return `[${visibleText}](${linkBaseUrl}${link})`;
+};
+
+const replaceWikiLinks = (body) => {
+  return body.replace(
+    /\[\[(.*?)\]\]/g,
+    (_, match) => getWikiLink(match)
+  );
+};
+
+const formatByline = byline => byline
+  .split(' and ')
+  .map((person) => '[[' + person.trim() + ']]')
+  .join(' and ');
+
+const formatSectionHeader = (item) => `
+
+# ${ item.title }
+
+_Section edited by ${item.byline && formatByline(item.byline)}._
+
+`;
+
+const addImageUrl = (item) => item.image ? `
+
+![Item cover image](${item.image})
+
+` : '';
+
+const formatItem = (item) => `${addImageUrl(item)}
+
+__${item.title}__ ${(item.body || '').replace(/\n/g, ' ')}
+
+`;
+
 const Markdown = ({ markdown, render }) => render ?
   (
     <div
@@ -27,13 +61,6 @@ const Markdown = ({ markdown, render }) => render ?
   ) : (
     <pre style={{ whiteSpace: 'pre-wrap' }}>{ markdown }</pre>
   );
-
-const Highlight = ({ item, render }) => (
-  <Markdown
-    render={ render }
-    markdown={ formatHighlights(item) }
-  />
-);
 
 const Section = ({ item, render }) => (
   <Markdown
@@ -121,16 +148,9 @@ const Zine = ({ content, error, renderMarkdown, loading, handleClipboard }) => {
         </Button>
       }
       <div id="content">
-        {content.map((item, i) => {
-          switch (item.section) {
-          case 'highlight':
-            return <Highlight key={`item-${i}`} render={renderMarkdown} item={item} />;
-          case 'heading':
-            return <Section key={`item-${i}`} render={renderMarkdown} item={item} />;
-          default:
-            return <Body key={`item-${i}`} render={renderMarkdown} item={item} />;
-          }
-        }
+        {content.map((item, i) => item.isSectionHeading
+          ? (<Section key={`item-${i}`} render={renderMarkdown} item={item} />)
+          : (<Body key={`item-${i}`} render={renderMarkdown} item={item} />)
         )}
       </div>
       {content.length === 0 && <em>No cards to show.</em>}
