+++ src/containers/MainContainer.js
@@ -5,7 +5,7 @@ import MainPage from 'components/MainPage';
 function mapStateToProps(state) {
   return {
     content: state.trello.get('content').toJS(),
-    error: state.trello.get('error'),
+    error: state.trello.get('error').toJS(),
     renderMarkdown: state.trello.get('renderMarkdown'),
     loading: state.trello.get('loading'),
   };
