+++ src/index.js
@@ -3,20 +3,22 @@ import 'styles/styles.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { Router, browserHistory } from 'react-router';
+import { syncReduxAndRouter } from 'redux-simple-router';
+import { Router, hashHistory } from 'react-router';
 import routes from './routes';
 import configureStore from './configureStore';
 
 import { fetchMyCards } from 'redux/trello/actions';
 
 const store = configureStore({});
+syncReduxAndRouter(hashHistory, store);
 
 store.dispatch(fetchMyCards());
 
 ReactDOM.render(
   <div>
     <Provider store={store}>
-      <Router history={browserHistory}>
+      <Router history={hashHistory}>
         {routes}
       </Router>
     </Provider>
