+++ src/services/api.js
@@ -2,62 +2,36 @@ import R from 'ramda';
 import Q from 'q';
 import Promise from 'bluebird';
 import moment from 'moment';
-import Rules from './rules';
-import {
-  addFormatting,
-  makeHighlights,
-  hasLabel,
-  makeSectionCard,
-} from 'utils/formatting/text';
+
+import Rules from 'services/rules';
 
 const get = (path) => new Promise((resolve, reject) => {
   Trello.get(path, resolve, reject);
 });
 
-const addAttachmentURLs = (card) => {
-  if (card.idAttachmentCover) {
-    return get(`/cards/${card.id}/attachments/${card.idAttachmentCover}`)
-      .then(attachment => R.merge(card, {
-        image: attachment.url,
-      }));
-  }
-  return Q.when(card);
-};
+const replaceDoubleQuotes = (string) => string.replace(
+  /\"([^\"]*?)\"/g,
+  (_, match) => `“${match}”`
+);
 
-const authorize = ({ onSuccess, onFailure }) => {
-  /* eslint no-console: 0 */
-  console.log('Authorizing trello');
-  Trello.authorize({
-    type: 'popup',
-    name: 'Newsletter Generator',
-    scope: {
-      read: true,
-      write: true,
-    },
-    expiration: 'never',
-    success: onSuccess,
-    error: onFailure,
-  });
-};
+const replaceApostrophes = (string) => string.replace(/\'/g, '’');
 
-const updateCard = (id) => {
-  const comment = {
-    text: `Validated on ${ moment().format('MMMM Do YYYY, h:mm:ss a') }`,
-  };
+const prepare = R.pipe(
+  replaceDoubleQuotes,
+  replaceApostrophes
+);
 
-  const success = successMsg => {
-    console.log('Posted comment to card', successMsg);
-  };
+const hasEndOfSentencePunctuation = (string) => '.!?'.indexOf(string.slice(-1)) === -1;
 
-  const error = errorMsg => {
-    console.log(`error: ${errorMsg}`);
-  };
-
-  Trello.post(`/cards/${id}/actions/comments/`, comment, success, error);
+const addPeriodIfMissing = (string) => {
+  const trimmed = string.trim();
+  return hasEndOfSentencePunctuation(trimmed) ?
+    trimmed + '.'
+    : trimmed;
 };
 
 const applyRules = (card) => {
-  if (!card.section) {
+  if (!card.isSectionHeading) {
     card.messages = [];
     const { list } = new Rules({});
 
@@ -93,15 +67,14 @@ const filterByEditor = async (lists) => {
   return lists.filter(list => new RegExp(editor.fullName).test(list.name));
 };
 
-const filterByMention = async (lists) => {
+const filterByMention = async (list) => {
   const user = await getAuthUser();
 
-  return Promise.filter(lists, async (card) => {
-    if (!card.section) {
+  return Promise.filter(list, async (card) => {
+    if (!card.isSectionHeading) {
       try {
         const actions = await getActions(card);
-        const match = actions.length > 0 && actions
-          .filter(action => action.idMemberCreator === user.id);
+        const match = actions.length > 0 && actions.filter(action => action.idMemberCreator === user.id);
 
         return match.length > 0;
       } catch (error) {
@@ -112,11 +85,48 @@ const filterByMention = async (lists) => {
   .then(filtered => filtered);
 };
 
+const addFormatting = (card) => R.merge(card, {
+  title: card.title || prepare(addPeriodIfMissing(card.name)),
+  body: card.desc && prepare(card.desc),
+});
+
+const addAttachmentURLs = (card) => {
+  if (card.idAttachmentCover) {
+    return get(`/cards/${card.id}/attachments/${card.idAttachmentCover}`)
+      .then(attachment => R.merge(card, {
+        image: attachment.url,
+      }));
+  }
+  return Q.when(card);
+};
+
+// const formatCard = (card) => card.isSectionHeading ?
+//   `## ${card.title}\n\n_Section edited by [[${card.byline}]]_.`
+//   : `__${prepare(addPeriodIfMissing(card.name))}__ ${prepare(card.desc)}`;
+
+// const log = (heading) => (item) => {
+//   /* eslint no-console: 0 */
+//   console.log(heading, item);
+//   return item;
+// };
+
+const makeSectionCard = (list) => ({
+  isSectionHeading: true,
+  title: list.name.split('/')[0],
+  byline: list.name.split('/')[1],
+});
+
+
 const getList = (list) => get(`/lists/${list.id}/cards`)
   .then((cards) => cards.length ?
     R.flatten([makeSectionCard(list), cards])
     : []
   );
+
+const hasLabel = (targetLabel) => (card) => R.contains(
+  targetLabel,
+  R.map((label) => label.name, card.labels || []), // all label names
+);
 
 const getMyCards = R.pipeP(
   (boardId) => get(`/boards/${boardId}/lists`),
@@ -126,7 +136,6 @@ const getMyCards = R.pipeP(
   R.flatten,
   R.reject(hasLabel('HOLD')),
   filterByMention,
-  makeHighlights,
   R.map(applyRules),
   R.map(addFormatting),
   R.map(addAttachmentURLs),
@@ -141,7 +150,6 @@ const getMySection = R.pipeP(
   (listOfPromises) => Promise.all(listOfPromises),
   R.flatten,
   R.reject(hasLabel('HOLD')),
-  makeHighlights,
   R.map(applyRules),
   R.map(addFormatting),
   R.map(addAttachmentURLs),
@@ -155,12 +163,44 @@ const getAllCards = R.pipeP(
   (listOfPromises) => Promise.all(listOfPromises),
   R.flatten,
   R.reject(hasLabel('HOLD')),
-  makeHighlights,
   R.map(applyRules),
   R.map(addFormatting),
   R.map(addAttachmentURLs),
   Q.all,
+  // (renderedCards) => renderedCards.join('\n\n'),
 );
+
+const authorize = ({ onSuccess, onFailure }) => {
+  /* eslint no-console: 0 */
+  console.log('Authorizing trello');
+  Trello.authorize({
+    type: 'popup',
+    name: 'Newsletter Generator',
+    scope: {
+      read: true,
+      write: true,
+    },
+    expiration: 'never',
+    success: onSuccess,
+    error: onFailure,
+  });
+};
+
+const updateCard = (id) => {
+  const comment = {
+    text: `Validated on ${ moment().format('MMMM Do YYYY, h:mm:ss a') }`,
+  };
+
+  const success = successMsg => {
+    console.log('Posted comment to card', successMsg);
+  };
+
+  const error = errorMsg => {
+    console.log(`error: ${errorMsg}`);
+  };
+
+  Trello.post(`/cards/${id}/actions/comments/`, comment, success, error);
+};
 
 export {
   getMyCards,
